Unity_VR/Backend/requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
